LibWeb: Split PositionStyleValue out of StyleValue.{h,cpp}
