+++ Userland/Libraries/LibWeb/CSS/StyleValue.h
@@ -627,39 +627,6 @@ private:
     NonnullOwnPtr<CalcSum> m_expression;
 };
 
-class PositionStyleValue final : public StyleValueWithDefaultOperators<PositionStyleValue> {
-public:
-    static ValueComparingNonnullRefPtr<PositionStyleValue> create(PositionEdge edge_x, LengthPercentage const& offset_x, PositionEdge edge_y, LengthPercentage const& offset_y)
-    {
-        return adopt_ref(*new PositionStyleValue(edge_x, offset_x, edge_y, offset_y));
-    }
-    virtual ~PositionStyleValue() override = default;
-
-    PositionEdge edge_x() const { return m_properties.edge_x; }
-    LengthPercentage const& offset_x() const { return m_properties.offset_x; }
-    PositionEdge edge_y() const { return m_properties.edge_y; }
-    LengthPercentage const& offset_y() const { return m_properties.offset_y; }
-
-    virtual ErrorOr<String> to_string() const override;
-
-    bool properties_equal(PositionStyleValue const& other) const { return m_properties == other.m_properties; }
-
-private:
-    PositionStyleValue(PositionEdge edge_x, LengthPercentage const& offset_x, PositionEdge edge_y, LengthPercentage const& offset_y)
-        : StyleValueWithDefaultOperators(Type::Position)
-        , m_properties { .edge_x = edge_x, .offset_x = offset_x, .edge_y = edge_y, .offset_y = offset_y }
-    {
-    }
-
-    struct Properties {
-        PositionEdge edge_x;
-        LengthPercentage offset_x;
-        PositionEdge edge_y;
-        LengthPercentage offset_y;
-        bool operator==(Properties const&) const = default;
-    } m_properties;
-};
-
 class ResolutionStyleValue : public StyleValueWithDefaultOperators<ResolutionStyleValue> {
 public:
     static ValueComparingNonnullRefPtr<ResolutionStyleValue> create(Resolution resolution)
